Actual mol writing now moved to MolvecOptions and improvements to Main for multithreaded SDF writing.
